# perf(remix-server-runtime): use faster alternative to `jsesc`