--- a/.changeset/four-numbers-end.md
+++ b/.changeset/four-numbers-end.md
@@ -1,0 +1,5 @@
+---
+"@remix-run/server-runtime": patch
+---
+
+Improve performance when serializing data in the server runtime.