--- a/contributors.yml
+++ b/contributors.yml
@@ -322,6 +322,7 @@
 - raulrpearson
 - real34
 - realjokele
+- redabacha
 - reggie3
 - rlfarman
 - roachjc