--- a/packages/remix-dev/package.json
+++ b/packages/remix-dev/package.json
@@ -64,6 +64,7 @@
     "@types/gunzip-maybe": "^1.4.0",
     "@types/inquirer": "^8.2.0",
     "@types/jscodeshift": "^0.11.3",
+    "@types/jsesc": "^3.0.1",
     "@types/lodash.debounce": "^4.0.6",
     "@types/npmcli__package-json": "^2.0.0",
     "@types/shelljs": "^0.8.11",