--- a/packages/remix-server-runtime/serverHandoff.ts
+++ b/packages/remix-server-runtime/serverHandoff.ts
@@ -1,7 +1,7 @@
-import jsesc from "jsesc";
+import { escapeHtml } from "./markup";
 
 export function createServerHandoffString(serverHandoff: any): string {
-  // Use jsesc to escape data returned from the loaders. This string is
-  // inserted directly into the HTML in the `<Scripts>` element.
-  return jsesc(serverHandoff, { isScriptContext: true });
+  // Uses faster alternative of jsesc to escape data returned from the loaders.
+  // This string is inserted directly into the HTML in the `<Scripts>` element.
+  return escapeHtml(JSON.stringify(serverHandoff));
 }