@jacob-ebey 2022-08-05
Going to have to double check, but I believe the serialization method needs to be consistent across machines / runtimes. I.e, `serialize(obj)` should return the same thing on the server and in the browser. `JSON.stringify` alone will not suffice here as it does not take into account key order during serialization leading to possible mismatchs on hydration.