@redabacha 2022-08-05
> Going to have to double check, but I believe the serialization method needs to be consistent across machines / runtimes. I.e, `serialize(obj)` should return the same thing on the server and in the browser. `JSON.stringify` alone will not suffice here as it does not take into account key order during serialization leading to possible mismatchs on hydration. Going to need something like [json-stable-stringify](https://github.com/substack/json-stable-stringify)

according to mdn, https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify#description:
> Enumerable properties are visited using the same algorithm as [Object.keys](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/keys), which has a well-defined order and is stable across implementations. For example, JSON.stringify on the same object will always produce the same string, and JSON.parse(JSON.stringify(obj)) would produce an object with the same key ordering as the original (assuming the object is completely JSON-serializable).

so my guess is that it would be fine. next.js seem to be getting away with just using [JSON.stringify](https://github.com/vercel/next.js/blob/34f5236f572b9f6a2e3c55a537a6a717ffaae71d/packages/next/pages/_document.tsx#L969-L1002) (and [JSON.parse on the client](https://github.com/vercel/next.js/blob/0f99768f9811fdf027a083471bd2a435f6207f4a/packages/next/client/index.tsx#L197-L199)).