@redabacha 2022-08-06
@jacob-ebey i've added additional benchmarks with [`json-stable-stringify`](https://www.npmjs.com/package/json-stable-stringify) and some other stable JSON.stringify packages i've found. according to my understanding though (see prior comment), they shouldn't be necessary in this day and age.